FinalProjectBE/requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
google-cloud-speech
pydub
python-multipart
noisereduce
scipy
numpy
